pyproject.toml
[project]
name = "parrotpark-sprig"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py>=2.1.0",
    "datasets==3.2.0",
    "dspy>=2.6.2",
    "immutabledict>=4.2.1",
    "langchain>=0.3.18",
    "langdetect>=1.0.9",
    "modal>=0.73.32",
    "nltk>=3.9.1",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "rouge-score>=0.1.2",
    "sacrebleu>=2.5.1",
    "scikit-learn>=1.6.1",
    "tqdm>=4.67.1",
]
